+++ CHANGES.rst
@@ -17,6 +17,7 @@ Development
 - Add basic device-interval sensors ``SystemTemperature`` and ``SystemBatteryLevel``
 - Explicitly shut down all peripherals having implicitly been turned on
 - Add ``SystemWiFiMetrics`` sensor
+- Add ``SystemUptime`` sensor
 
 2019-06-07 0.3.0
 ================